Change gain interpretation
The hardware uses a 0-31 attenuator setting (first 10 steps in
increments of 1dB, remaining in increments of 2dB, i.e. in
particular the precision is concentrated at the high end of
the amplifier). However, this appears to be primarily an
implementation detail, since the hardware block does act
as an amplifier (I think). In particular LimeSuite also
treats this as a 0-52dB amplifier, rather than an attenuator,
so let's be consistent here.